requirements.txt
flask
pyjwt
python-dotenv
requests